Add Manager project and update data handling logic
Added a new Python project, "Manager", to the solution file
with corresponding build configurations. Updated data type
handling in `FlightSimConnector.cs` from `FLOAT32` to
`FLOAT64` and improved logging for requests. Modified
`SimVarData.cs` to use `FLOAT64` as the default data type.
Ensured proper initialization in `DisplayLogging.cs` by
adding a call to `InitializeComponent()`.

+++ Classes/SimVarData.cs
@@ -18,7 +18,7 @@ public class MeasurementUnits
             {
                 Name = name,
                 Measure = "",
-                Type = "FLOAT32"
+                Type = "FLOAT64"
             };
         }
 
@@ -28,7 +28,7 @@ public class MeasurementUnits
             {
                 Name = name,
                 Measure = unit,
-                Type = "FLOAT32"
+                Type = "FLOAT64"
             };
         }
